added new counting clicks
Added a click counter using after. It turned out to be longer, not so universal, more complicated, but now everything can be configured more flexibly, like a timer waiting for a key press to change a letter.

But there is still one problem - since the timer is started and cancelled either (which is logical) at its end, or if the user presses the button more than necessary, if you enter not the last character, the timer will not end, so there is a VISUAL bug when you start to quickly enter the next letter: it changes to the previous one.
How to eliminate this is not very clear yet, I'm trying in counting.py, but it's not clear how to determine when the user has stopped in the selection of letters, and not thinking before another click. Now I'm trying to create a global variable, namely a flag that would forcibly close the previous timer